(db/snapshot): Add DropIndex API UT
