+++ core/src/db/SSDBImpl.h
@@ -74,8 +74,7 @@ class SSDBImpl {
     ShowPartitions(const std::string& collection_name, std::vector<std::string>& partition_names);
 
     Status
-    DropIndex(const std::string& collection_name, const std::string& field_name,
-            const std::string& field_element_name);
+    DropIndex(const std::string& collection_name, const std::string& field_name, const std::string& field_element_name);
 
  private:
     void
