+++ core/unittest/ssdb/test_db.cpp
@@ -18,55 +18,6 @@
 #include <algorithm>
 
 #include "ssdb/utils.h"
-#include "db/snapshot/CompoundOperations.h"
-#include "db/snapshot/Context.h"
-#include "db/snapshot/EventExecutor.h"
-#include "db/snapshot/OperationExecutor.h"
-#include "db/snapshot/ReferenceProxy.h"
-#include "db/snapshot/ResourceHolders.h"
-#include "db/snapshot/ScopedResource.h"
-#include "db/snapshot/Snapshots.h"
-#include "db/snapshot/Store.h"
-#include "db/snapshot/WrappedTypes.h"
-
-using ID_TYPE = milvus::engine::snapshot::ID_TYPE;
-using IDS_TYPE = milvus::engine::snapshot::IDS_TYPE;
-using LSN_TYPE = milvus::engine::snapshot::LSN_TYPE;
-using MappingT = milvus::engine::snapshot::MappingT;
-using LoadOperationContext = milvus::engine::snapshot::LoadOperationContext;
-using CreateCollectionContext = milvus::engine::snapshot::CreateCollectionContext;
-using SegmentFileContext = milvus::engine::snapshot::SegmentFileContext;
-using OperationContext = milvus::engine::snapshot::OperationContext;
-using PartitionContext = milvus::engine::snapshot::PartitionContext;
-using BuildOperation = milvus::engine::snapshot::BuildOperation;
-using MergeOperation = milvus::engine::snapshot::MergeOperation;
-using CreateCollectionOperation = milvus::engine::snapshot::CreateCollectionOperation;
-using NewSegmentOperation = milvus::engine::snapshot::NewSegmentOperation;
-using DropPartitionOperation = milvus::engine::snapshot::DropPartitionOperation;
-using CreatePartitionOperation = milvus::engine::snapshot::CreatePartitionOperation;
-using DropCollectionOperation = milvus::engine::snapshot::DropCollectionOperation;
-using CollectionCommitsHolder = milvus::engine::snapshot::CollectionCommitsHolder;
-using CollectionsHolder = milvus::engine::snapshot::CollectionsHolder;
-using CollectionScopedT = milvus::engine::snapshot::CollectionScopedT;
-using Collection = milvus::engine::snapshot::Collection;
-using CollectionPtr = milvus::engine::snapshot::CollectionPtr;
-using Partition = milvus::engine::snapshot::Partition;
-using PartitionPtr = milvus::engine::snapshot::PartitionPtr;
-using Segment = milvus::engine::snapshot::Segment;
-using SegmentPtr = milvus::engine::snapshot::SegmentPtr;
-using SegmentFile = milvus::engine::snapshot::SegmentFile;
-using SegmentFilePtr = milvus::engine::snapshot::SegmentFilePtr;
-using Field = milvus::engine::snapshot::Field;
-using FieldElement = milvus::engine::snapshot::FieldElement;
-using Snapshots = milvus::engine::snapshot::Snapshots;
-using ScopedSnapshotT = milvus::engine::snapshot::ScopedSnapshotT;
-using ReferenceProxy = milvus::engine::snapshot::ReferenceProxy;
-using Queue = milvus::BlockingQueue<ID_TYPE>;
-using TQueue = milvus::BlockingQueue<std::tuple<ID_TYPE, ID_TYPE>>;
-using SoftDeleteCollectionOperation = milvus::engine::snapshot::SoftDeleteOperation<Collection>;
-using ParamsField = milvus::engine::snapshot::ParamsField;
-using IteratePartitionHandler = milvus::engine::snapshot::IterateHandler<Partition>;
-using SSDBImpl = milvus::engine::SSDBImpl;
 
 milvus::Status
 CreateCollection(std::shared_ptr<SSDBImpl> db, const std::string& collection_name, const LSN_TYPE& lsn) {
@@ -179,4 +130,77 @@ TEST_F(SSDBTest, PartitionTest) {
     status = db_->ShowPartitions(c1, partition_names);
     ASSERT_TRUE(status.ok());
     ASSERT_EQ(partition_names.size(), 1);
+}
+
+TEST_F(SSDBTest, IndexTest) {
+    LSN_TYPE lsn = 0;
+    auto next_lsn = [&]() -> decltype(lsn) {
+        return ++lsn;
+    };
+
+    std::string c1 = "c1";
+    auto status = CreateCollection(db_, c1, next_lsn());
+    ASSERT_TRUE(status.ok());
+
+    std::stringstream p_name;
+    auto num = RandomInt(3, 5);
+    for (auto i = 0; i < num; ++i) {
+        p_name.str("");
+        p_name << "partition_" << i;
+        status = db_->CreatePartition(c1, p_name.str());
+        ASSERT_TRUE(status.ok());
+    }
+
+    ScopedSnapshotT ss;
+    status = Snapshots::GetInstance().GetSnapshot(ss, c1);
+    ASSERT_TRUE(status.ok());
+
+    SegmentFileContext sf_context;
+    SFContextBuilder(sf_context, ss);
+
+    auto new_total = 0;
+    auto& partitions = ss->GetResources<Partition>();
+    for (auto& kv : partitions) {
+        num = RandomInt(2, 5);
+        for (auto i = 0; i < num; ++i) {
+            ASSERT_TRUE(CreateSegment(ss, kv.first, next_lsn(), sf_context).ok());
+        }
+        new_total += num;
+    }
+
+    auto field_element_id = ss->GetFieldElementId(sf_context.field_name, sf_context.field_element_name);
+    ASSERT_NE(field_element_id, 0);
+
+    auto filter1 = [&](SegmentFile::Ptr segment_file) -> bool {
+        if (segment_file->GetFieldElementId() == field_element_id) {
+            return true;
+        }
+        return false;
+    };
+
+    status = Snapshots::GetInstance().GetSnapshot(ss, c1);
+    ASSERT_TRUE(status.ok());
+    auto sf_collector = std::make_shared<SegmentFileCollector>(ss, filter1);
+    sf_collector->Iterate();
+    std::cout << "Total " << sf_collector->segment_files_.size() << " of field_element_id ";
+    std::cout << field_element_id << std::endl;
+    ASSERT_EQ(new_total, sf_collector->segment_files_.size());
+
+    status = db_->DropIndex(c1, sf_context.field_name, sf_context.field_element_name);
+    ASSERT_TRUE(status.ok());
+
+    status = Snapshots::GetInstance().GetSnapshot(ss, c1);
+    ASSERT_TRUE(status.ok());
+    sf_collector = std::make_shared<SegmentFileCollector>(ss, filter1);
+    sf_collector->Iterate();
+    std::cout << "Total " << sf_collector->segment_files_.size() << " of field_element_id ";
+    std::cout << field_element_id << std::endl;
+    ASSERT_EQ(0, sf_collector->segment_files_.size());
+
+    {
+        auto& field_elements = ss->GetResources<FieldElement>();
+        for (auto& kv : field_elements) {
+            ASSERT_NE(kv.second->GetID(), field_element_id);
+        }
+    }
 }
