+++ core/unittest/ssdb/test_snapshot.cpp
@@ -18,195 +18,6 @@
 #include <algorithm>
 
 #include "ssdb/utils.h"
-#include "db/snapshot/CompoundOperations.h"
-#include "db/snapshot/Context.h"
-#include "db/snapshot/EventExecutor.h"
-#include "db/snapshot/OperationExecutor.h"
-#include "db/snapshot/ReferenceProxy.h"
-#include "db/snapshot/ResourceHolders.h"
-#include "db/snapshot/ScopedResource.h"
-#include "db/snapshot/Snapshots.h"
-#include "db/snapshot/Store.h"
-#include "db/snapshot/WrappedTypes.h"
-
-using ID_TYPE = milvus::engine::snapshot::ID_TYPE;
-using IDS_TYPE = milvus::engine::snapshot::IDS_TYPE;
-using LSN_TYPE = milvus::engine::snapshot::LSN_TYPE;
-using MappingT = milvus::engine::snapshot::MappingT;
-using LoadOperationContext = milvus::engine::snapshot::LoadOperationContext;
-using CreateCollectionContext = milvus::engine::snapshot::CreateCollectionContext;
-using SegmentFileContext = milvus::engine::snapshot::SegmentFileContext;
-using OperationContext = milvus::engine::snapshot::OperationContext;
-using PartitionContext = milvus::engine::snapshot::PartitionContext;
-using DropIndexOperation = milvus::engine::snapshot::DropIndexOperation;
-using DropAllIndexOperation = milvus::engine::snapshot::DropAllIndexOperation;
-using BuildOperation = milvus::engine::snapshot::BuildOperation;
-using MergeOperation = milvus::engine::snapshot::MergeOperation;
-using CreateCollectionOperation = milvus::engine::snapshot::CreateCollectionOperation;
-using NewSegmentOperation = milvus::engine::snapshot::NewSegmentOperation;
-using DropPartitionOperation = milvus::engine::snapshot::DropPartitionOperation;
-using CreatePartitionOperation = milvus::engine::snapshot::CreatePartitionOperation;
-using DropCollectionOperation = milvus::engine::snapshot::DropCollectionOperation;
-using CollectionCommitsHolder = milvus::engine::snapshot::CollectionCommitsHolder;
-using CollectionsHolder = milvus::engine::snapshot::CollectionsHolder;
-using CollectionScopedT = milvus::engine::snapshot::CollectionScopedT;
-using Collection = milvus::engine::snapshot::Collection;
-using CollectionPtr = milvus::engine::snapshot::CollectionPtr;
-using Partition = milvus::engine::snapshot::Partition;
-using PartitionPtr = milvus::engine::snapshot::PartitionPtr;
-using Segment = milvus::engine::snapshot::Segment;
-using SegmentPtr = milvus::engine::snapshot::SegmentPtr;
-using SegmentFile = milvus::engine::snapshot::SegmentFile;
-using SegmentFilePtr = milvus::engine::snapshot::SegmentFilePtr;
-using Field = milvus::engine::snapshot::Field;
-using FieldElement = milvus::engine::snapshot::FieldElement;
-using Snapshots = milvus::engine::snapshot::Snapshots;
-using ScopedSnapshotT = milvus::engine::snapshot::ScopedSnapshotT;
-using ReferenceProxy = milvus::engine::snapshot::ReferenceProxy;
-using Queue = milvus::BlockingQueue<ID_TYPE>;
-using TQueue = milvus::BlockingQueue<std::tuple<ID_TYPE, ID_TYPE>>;
-using SoftDeleteCollectionOperation = milvus::engine::snapshot::SoftDeleteOperation<Collection>;
-using ParamsField = milvus::engine::snapshot::ParamsField;
-using IteratePartitionHandler = milvus::engine::snapshot::IterateHandler<Partition>;
-using IterateSegmentFileHandler = milvus::engine::snapshot::IterateHandler<SegmentFile>;
-
-void SFContextBuilder(SegmentFileContext& ctx, ScopedSnapshotT sss) {
-    auto field = sss->GetResources<Field>().begin()->second;
-    ctx.field_name = field->GetName();
-    for (auto& kv : sss->GetResources<FieldElement>()) {
-        ctx.field_element_name = kv.second->GetName();
-        break;
-    }
-
-    ctx.segment_id = sss->GetResources<Segment>().begin()->second->GetID();
-    ctx.partition_id = sss->GetResources<Segment>().begin()->second->GetPartitionId();
-}
-
-struct PartitionCollector : public IteratePartitionHandler {
-    using ResourceT = Partition;
-    using BaseT = IteratePartitionHandler;
-    explicit PartitionCollector(ScopedSnapshotT ss) : BaseT(ss) {}
-
-    milvus::Status
-    PreIterate() override {
-        partition_names_.clear();
-        return milvus::Status::OK();
-    }
-
-    milvus::Status
-    Handle(const typename ResourceT::Ptr& partition) override {
-        partition_names_.push_back(partition->GetName());
-        return milvus::Status::OK();
-    }
-
-    std::vector<std::string> partition_names_;
-};
-
-using FilterT = std::function<bool(SegmentFile::Ptr)>;
-struct SegmentFileCollector : public IterateSegmentFileHandler {
-    using ResourceT = SegmentFile;
-    using BaseT = IterateSegmentFileHandler;
-    explicit SegmentFileCollector(ScopedSnapshotT ss, const FilterT& filter)
-        : filter_(filter), BaseT(ss) {}
-
-    milvus::Status
-    PreIterate() override {
-        segment_files_.clear();
-        return milvus::Status::OK();
-    }
-
-    milvus::Status
-    Handle(const typename ResourceT::Ptr& segment_file) override {
-        if (!filter_(segment_file)) {
-            return milvus::Status::OK();
-        }
-        segment_files_.insert(segment_file->GetID());
-        return milvus::Status::OK();
-    }
-
-    FilterT filter_;
-    std::set<ID_TYPE> segment_files_;
-};
-
-struct WaitableObj {
-    bool notified_ = false;
-    std::mutex mutex_;
-    std::condition_variable cv_;
-
-    void
-    Wait() {
-        std::unique_lock<std::mutex> lck(mutex_);
-        if (!notified_) {
-            cv_.wait(lck);
-        }
-        notified_ = false;
-    }
-
-    void
-    Notify() {
-        std::unique_lock<std::mutex> lck(mutex_);
-        notified_ = true;
-        lck.unlock();
-        cv_.notify_one();
-    }
-};
-
-ScopedSnapshotT
-CreateCollection(const std::string& collection_name, const LSN_TYPE& lsn) {
-    CreateCollectionContext context;
-    context.lsn = lsn;
-    auto collection_schema = std::make_shared<Collection>(collection_name);
-    context.collection = collection_schema;
-    auto vector_field = std::make_shared<Field>("vector", 0,
-            milvus::engine::snapshot::FieldType::VECTOR);
-    auto vector_field_element = std::make_shared<FieldElement>(0, 0, "ivfsq8",
-            milvus::engine::snapshot::FieldElementType::IVFSQ8);
-    auto int_field = std::make_shared<Field>("int", 0,
-            milvus::engine::snapshot::FieldType::INT32);
-    context.fields_schema[vector_field] = {vector_field_element};
-    context.fields_schema[int_field] = {};
-
-    auto op = std::make_shared<CreateCollectionOperation>(context);
-    op->Push();
-    ScopedSnapshotT ss;
-    auto status = op->GetSnapshot(ss);
-    return ss;
-}
-
-ScopedSnapshotT
-CreatePartition(const std::string& collection_name, const PartitionContext& p_context, const LSN_TYPE& lsn) {
-    ScopedSnapshotT curr_ss;
-    ScopedSnapshotT ss;
-    auto status = Snapshots::GetInstance().GetSnapshot(ss, collection_name);
-    if (!status.ok()) {
-        std::cout << status.ToString() << std::endl;
-        return curr_ss;
-    }
-
-    OperationContext context;
-    context.lsn = lsn;
-    auto op = std::make_shared<CreatePartitionOperation>(context, ss);
-
-    PartitionPtr partition;
-    status = op->CommitNewPartition(p_context, partition);
-    if (!status.ok()) {
-        std::cout << status.ToString() << std::endl;
-        return curr_ss;
-    }
-
-    status = op->Push();
-    if (!status.ok()) {
-        std::cout << status.ToString() << std::endl;
-        return curr_ss;
-    }
-
-    status = op->GetSnapshot(curr_ss);
-    if (!status.ok()) {
-        std::cout << status.ToString() << std::endl;
-        return curr_ss;
-    }
-    return curr_ss;
-}
 
 TEST_F(SnapshotTest, ResourcesTest) {
     int nprobe = 16;
@@ -464,7 +275,7 @@ TEST_F(SnapshotTest, ConCurrentCollectionOperation) {
 
     ID_TYPE stale_ss_id;
     auto worker1 = [&]() {
-        milvus::Status status;
+        Status status;
         auto ss = CreateCollection(collection_name, ++lsn);
         ASSERT_TRUE(ss);
         ASSERT_EQ(ss->GetName(), collection_name);
@@ -714,37 +525,18 @@ TEST_F(SnapshotTest, IndexTest) {
     sf_collector->Iterate();
     auto prev_total = sf_collector->segment_files_.size();
 
-    auto create_segment = [&](ID_TYPE partition_id) {
-        OperationContext context;
-        context.lsn = next_lsn();
-        context.prev_partition = ss->GetResource<Partition>(partition_id);
-        auto op = std::make_shared<NewSegmentOperation>(context, ss);
-        SegmentPtr new_seg;
-        status = op->CommitNewSegment(new_seg);
-        ASSERT_TRUE(status.ok());
-        ASSERT_FALSE(new_seg->ToString().empty());
-        SegmentFilePtr seg_file;
-        auto nsf_context = sf_context;
-        nsf_context.segment_id = new_seg->GetID();
-        nsf_context.partition_id = new_seg->GetPartitionId();
-        status = op->CommitNewSegmentFile(nsf_context, seg_file);
-        ASSERT_TRUE(status.ok());
-        status = op->Push();
-        ASSERT_TRUE(status.ok());
-
-        status = op->GetSnapshot(ss);
-        ASSERT_TRUE(status.ok());
-    };
-
     auto new_total = 0;
     auto partitions = ss->GetResources<Partition>();
     for (auto& kv : partitions) {
         num = RandomInt(2, 5);
         for (auto i = 0; i < num; ++i) {
-            create_segment(kv.first);
+            ASSERT_TRUE(CreateSegment(ss, kv.first, next_lsn(), sf_context).ok());
         }
         new_total += num;
     }
+
+    status = Snapshots::GetInstance().GetSnapshot(ss, ss->GetName());
+    ASSERT_TRUE(status.ok());
 
     sf_collector = std::make_shared<SegmentFileCollector>(ss, filter2);
     sf_collector->Iterate();
@@ -804,7 +596,7 @@ TEST_F(SnapshotTest, IndexTest) {
 }
 
 TEST_F(SnapshotTest, OperationTest) {
-    milvus::Status status;
+    Status status;
     std::string to_string;
     LSN_TYPE lsn;
 
@@ -981,7 +773,7 @@ TEST_F(SnapshotTest, OperationTest) {
 }
 
 TEST_F(SnapshotTest, CompoundTest1) {
-    milvus::Status status;
+    Status status;
     std::atomic<LSN_TYPE> lsn = 0;
     auto next_lsn = [&]() -> decltype(lsn) {
         return ++lsn;
@@ -1322,7 +1114,7 @@ TEST_F(SnapshotTest, CompoundTest1) {
 
 
 TEST_F(SnapshotTest, CompoundTest2) {
-    milvus::Status status;
+    Status status;
     LSN_TYPE lsn = 0;
     auto next_lsn = [&]() -> LSN_TYPE& {
         return ++lsn;
